% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RCB2__AssetCodeRCB2}
\alias{RCB2__AssetCodeRCB2}
\title{Dictionary for AssetCodeRCB2 variable of RCB2 file}
\format{
\subsection{\code{RCB2__AssetCodeRCB2}}{

A data frame with 17 rows and 2 columns:
\describe{
\item{code}{Integer code}
\item{value}{Character description}
}
}
}
\source{
D_RCB2.TXT
}
\usage{
RCB2__AssetCodeRCB2
}
\description{
Dictionary for AssetCodeRCB2 variable of RCB2 file
}
\keyword{datasets}
