% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RCB3__DebtMaturityCode}
\alias{RCB3__DebtMaturityCode}
\title{Dictionary for DebtMaturityCode variable of RCB3 file}
\format{
\subsection{\code{RCB3__DebtMaturityCode}}{

A data frame with 10 rows and 2 columns:
\describe{
\item{code}{Integer code}
\item{value}{Character description}
}
}
}
\source{
D_RCB3.TXT
}
\usage{
RCB3__DebtMaturityCode
}
\description{
Dictionary for DebtMaturityCode variable of RCB3 file
}
\keyword{datasets}
