% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RCF__LOANSTATUS}
\alias{RCF__LOANSTATUS}
\title{Dictionary for LOANSTATUS variable of RCF file}
\format{
\subsection{\code{RCF__LOANSTATUS}}{

A data frame with 6 rows and 2 columns:
\describe{
\item{code}{Integer code}
\item{value}{Character description}
}
}
}
\source{
D_RCF.TXT
}
\usage{
RCF__LOANSTATUS
}
\description{
Dictionary for LOANSTATUS variable of RCF file
}
\keyword{datasets}
