% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RCI2C__ExposureCode}
\alias{RCI2C__ExposureCode}
\title{Dictionary for ExposureCode variable of RCI2C file}
\format{
\subsection{\code{RCI2C__ExposureCode}}{

A data frame with 12 rows and 2 columns:
\describe{
\item{code}{Integer code}
\item{value}{Character description}
}
}
}
\source{
D_RCI2C_2018.TXT
}
\usage{
RCI2C__ExposureCode
}
\description{
Dictionary for ExposureCode variable of RCI2C file
}
\keyword{datasets}
