% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RCR7__RegCapCode}
\alias{RCR7__RegCapCode}
\title{Dictionary for RegCapCode variable of RCR7 file}
\format{
\subsection{\code{RCR7__RegCapCode}}{

A data frame with 29 rows and 2 columns:
\describe{
\item{code}{Integer code}
\item{value}{Character description}
}
}
}
\source{
D_RCR7.TXT
}
\usage{
RCR7__RegCapCode
}
\description{
Dictionary for RegCapCode variable of RCR7 file
}
\keyword{datasets}
