% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RIE1__ACLCode}
\alias{RIE1__ACLCode}
\title{Dictionary for ACLCode variable of RIE1 file}
\format{
\subsection{\code{RIE1__ACLCode}}{

A data frame with 7 rows and 2 columns:
\describe{
\item{code}{Integer code}
\item{value}{Character description}
}
}
}
\source{
D_RIE1.TXT
}
\usage{
RIE1__ACLCode
}
\description{
Dictionary for ACLCode variable of RIE1 file
}
\keyword{datasets}
