% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_metadata.R
\name{compare_metadata}
\alias{compare_metadata}
\title{Compare metadata files between two folders}
\usage{
compare_metadata(folder1, folder2)
}
\arguments{
\item{folder1}{A character string specifying the path to the first folder.}

\item{folder2}{A character string specifying the path to the second folder.}
}
\value{
A list containing information about differences in file names, file order,
and content differences for shared metadata files.
}
\description{
This function compares metadata files between two specified folders. It identifies
differences in the number of files, file names, and their order. Additionally, it
compares the content of shared metadata files to identify any differences.
}
\details{
The function lists metadata files in each folder, identifies shared metadata
files, and then compares the number of files, file names, and their order using the
\code{waldo::compare} function. It further compares the content of shared metadata
files using the \code{compare_files_content} function.
}
