% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{download_data}
\alias{download_data}
\title{Download data from FCA website}
\usage{
download_data(year, month, dest, files = NULL)
}
\arguments{
\item{year}{Integer, the year of the call report (e.g., 2022)}

\item{month}{Integer, they month of the call report (e.g., 3)}

\item{dest}{String, representing the directory where the data will be
downloaded into}

\item{files}{(Optional) character vector, representing the names of the files
in the .zip archive to be downloaded. If left \code{NULL}, all files will be
downloaded.}
}
\value{
Console message informing the user where the data was successfully
downloaded to
}
\description{
Download data from FCA website
}
\details{
FCA publishes call reports are quarterly, and are typically dated
as "March", "June", "September", and "December". Therefore, valid values
to the \code{month} argument should be limited to \code{c(3, 6, 9, 12)}, unless there
is an anomoly in FCA's reporting/publishing. Check
\url{https://www.fca.gov/bank-oversight/call-report-data-for-download} to
ensure the data is available for the quarter you are interested in.
}
\examples{
\dontrun{
  download_data(
    year = 2022,
    month = 12,
    dest = "data",
    c("D_INST.TXT", "D_RC.TXT")
   )
}
}
