% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{file_metadata}
\alias{file_metadata}
\title{Descriptions for data files}
\format{
\subsection{\code{file_metadata}}{

A data frame with 36 rows and 2 columns:
\describe{
\item{file_prefix}{Data file root name}
\item{description}{Short description of data file}
}
}
}
\source{
Metadata files headers
}
\usage{
file_metadata
}
\description{
Descriptions for data files
}
\keyword{datasets}
