% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{get_codes_dict}
\alias{get_codes_dict}
\title{Retrieve codes dictionary for a specified data name}
\usage{
get_codes_dict(data_name)
}
\arguments{
\item{data_name}{A character string specifying the data name to retrieve the
codes dictionary for.}
}
\value{
A list with the codes dictionary (\code{codes_dict}) and the associated
variable name (\code{codes_varname}) if found, otherwise each element will be NULL.
}
\description{
This function searches for an internal .rda file in the specified package
and retrieves the codes dictionary based on the provided data name and naming
convention. The naming convention is assumed to include the data name followed
by a double underscore "__". If the codes dictionary is found, it is returned;
otherwise, a NULL value is returned.
}
\details{
The function uses the provided data name to construct the expected
naming convention and searches for an internal .rda file in the specified package.
If found, it attempts to retrieve the codes dictionary using \code{get} and returns
it; otherwise, it returns NULL.
}
