% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{process_data}
\alias{process_data}
\title{Process data in a specified folder}
\usage{
process_data(folder)
}
\arguments{
\item{folder}{A character string specifying the path to the folder containing
the data and metadata files.}
}
\value{
A list containing processed data and metadata.
}
\description{
This function processes data files and their corresponding metadata files
in a given folder. It extracts information from metadata files (prefixed with
"D_") and applies the extracted metadata to process the associated data files.
}
\details{
The function assumes that metadata and data files share a common
root name (characters until the first underscore occurrence) and that
metadata and data files are sorted in the same order.
}
