% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RCB__INV_CODE}
\alias{RCB__INV_CODE}
\title{Dictionary for INV_CODE variable of RCB file}
\format{
\subsection{\code{RCB__INV_CODE}}{

A data frame with 35 rows and 2 columns:
\describe{
\item{code}{Integer code}
\item{value}{Character description}
}
}
}
\source{
D_RCB.TXT
}
\usage{
RCB__INV_CODE
}
\description{
Dictionary for INV_CODE variable of RCB file
}
\keyword{datasets}
