% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RCF1__LOANSTATUS}
\alias{RCF1__LOANSTATUS}
\title{Dictionary for LOANSTATUS variable of RCF1 file}
\format{
\subsection{\code{RCF1__LOANSTATUS}}{

A data frame with 13 rows and 2 columns:
\describe{
\item{code}{Integer code}
\item{value}{Character description}
}
}
}
\source{
D_RCF1.TXT
}
\usage{
RCF1__LOANSTATUS
}
\description{
Dictionary for LOANSTATUS variable of RCF1 file
}
\keyword{datasets}
