% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RCI2B__DerivCode}
\alias{RCI2B__DerivCode}
\title{Dictionary for DerivCode variable of RCI2B file}
\format{
\subsection{\code{RCI2B__DerivCode}}{

A data frame with 23 rows and 2 columns:
\describe{
\item{code}{Integer code}
\item{value}{Character description}
}
}
}
\source{
D_RCI2B_2018.TXT
}
\usage{
RCI2B__DerivCode
}
\description{
Dictionary for DerivCode variable of RCI2B file
}
\keyword{datasets}
