% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RCI2D__DerivRMCode}
\alias{RCI2D__DerivRMCode}
\title{Dictionary for DerivRMCode variable of RCI2D file}
\format{
\subsection{\code{RCI2D__DerivRMCode}}{

A data frame with 11 rows and 2 columns:
\describe{
\item{code}{Integer code}
\item{value}{Character description}
}
}
}
\source{
D_RCI2D_2018.TXT
}
\usage{
RCI2D__DerivRMCode
}
\description{
Dictionary for DerivRMCode variable of RCI2D file
}
\keyword{datasets}
