% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RCO__ASSET_CODE}
\alias{RCO__ASSET_CODE}
\title{Dictionary for ASSET_CODE variable of RCO file}
\format{
\subsection{\code{RCO__ASSET_CODE}}{

A data frame with 12 rows and 2 columns:
\describe{
\item{code}{Integer code}
\item{value}{Character description}
}
}
}
\source{
D_RCO.TXT
}
\usage{
RCO__ASSET_CODE
}
\description{
Dictionary for ASSET_CODE variable of RCO file
}
\keyword{datasets}
