% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RCR3__RegCapCode}
\alias{RCR3__RegCapCode}
\title{Dictionary for RegCapCode variable of RCR3 file}
\format{
\subsection{\code{RCR3__RegCapCode}}{

A data frame with 15 rows and 2 columns:
\describe{
\item{code}{Integer code}
\item{value}{Character description}
}
}
}
\source{
D_RCR3.TXT
}
\usage{
RCR3__RegCapCode
}
\description{
Dictionary for RegCapCode variable of RCR3 file
}
\keyword{datasets}
