% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{read_data_file}
\alias{read_data_file}
\title{Read and process a data file based on metadata and codes dictionary}
\usage{
read_data_file(filepath, metadata, codes_dict)
}
\arguments{
\item{filepath}{A character string specifying the path to the data file.}

\item{metadata}{A list containing the scenario and variable information obtained
from the metadata file using \code{\link{process_metadata_file}}.}

\item{codes_dict}{A data frame containing codes dictionary information.}
}
\value{
A tibble containing the processed data.
}
\description{
This function reads a data file and processes it based on the provided metadata
and codes dictionary. The processing depends on the metadata scenario, which
includes cases like "single," "single_multiple," and "single_multiple_single."
For certain scenarios, the function utilizes \code{read.csv} to infer column
types without explicit specification.
}
\details{
The function reads the data file and applies necessary processing based
on the metadata scenario. For scenarios like "single" and "single_multiple," it
uses \code{read.csv} for convenient type inference. For "single_multiple_single,"
it reads the file line by line, collapses every (N_CODES + 2) lines, and then reads
the collapsed lines using \code{read.table}.
}
