% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RID__CAP_CODE}
\alias{RID__CAP_CODE}
\title{Dictionary for RegCapCode variable of RID file}
\format{
\subsection{\code{RID__CAP_CODE}}{

A data frame with 12 rows and 2 columns:
\describe{
\item{code}{Integer code}
\item{value}{Character description}
}
}
}
\source{
D_RID.TXT
}
\usage{
RID__CAP_CODE
}
\description{
Dictionary for RegCapCode variable of RID file
}
\keyword{datasets}
